modified:   source/LOTRFellowshipoftheRingWSFix/LOTRFellowshipoftheRingWidescreenFix.cpp

+++ source/LOTRFellowshipoftheRingWSFix/LOTRFellowshipoftheRingWidescreenFix.cpp
@@ -67,7 +67,7 @@ void HandleChoiceInput(int &choice)
 }
 
 // Function to handle user input in resolution
-double HandleResolutionInput()
+int16_t HandleResolutionInput()
 {
     do
     {
